Update to prevent weird (Git? VS Code? Not sure what the source is) error
Quick update to prevent capitalization from being what causes this to not work at all....

Why was the name of the file flip flopping? There was no reason for it to.

Also, Git didn't find the new files after I renamed them to fix this issue. These are both issues I can't deal with, so I just made the code figure it out no matter which way it is capitalized.